requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.2.0
django-filter==23.5
stripe==7.11.0
django-environ==0.11.2
django-storages==1.14.2
django-crispy-forms==2.1
crispy-bootstrap5==2023.10 